[X86] Change target of __builtin_ia32_cmp[p|s][s|d] from avx into sse/sse2
